+++ packages/stream_chat_flutter_core/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming 
+
+- Deprecated the `sort` parameter in the `StreamChannelListController` in favor of `channelStateSort`.
+
 ## 5.0.0
 
 - Included the changes from version [4.5.0](#450).